✨ introduce ContextAnalyzer and ContextualSearchEngine for enhanced text analysis
- Implement ContextAnalyzer to classify text into emotional, temporal, and thematic contexts.
- Add ContextualSearchEngine to perform contextual searches based on user queries and mood data.
- Refactor EnhancedEmbeddingService to utilize new classes for context determination and search functionality.
- Introduce TemporalUtils and VectorUtils for date extraction and vector operations, respectively.
- Enhance test coverage by integrating ContextAnalyzer and VectorUtils into existing tests.
