+++ services/ai/EnhancedEmbeddingService.ts
@@ -2,6 +2,10 @@ import { App, TFile } from "obsidian";
 import { embed } from "../../ai/EmbeddingClient";
 import type { NovaJournalSettings } from "../../settings/PluginSettings";
 import type { MoodData } from "../rendering/FrontmatterService";
+import { ContextAnalyzer } from './ContextAnalyzer';
+import { TemporalUtils } from './TemporalUtils';
+import { VectorUtils } from './VectorUtils';
+import { ContextualSearchEngine } from './ContextualSearchEngine';
 
 export type ContextType = 'emotional' | 'temporal' | 'thematic' | 'general';
 
@@ -38,16 +42,18 @@ export interface SearchOptions {
 export class EnhancedEmbeddingService {
   private readonly indexPath = "nova-journal-enhanced-index.json";
   private readonly maxDays = 180;
-  private readonly chunkSize = 250;
-  private readonly overlap = 75;
   private readonly maxChunksPerBatch = 50;
   private readonly version = "2.0.0";
   private index: EnhancedIndexData | null = null;
+  private readonly contextAnalyzer = new ContextAnalyzer();
+  private readonly searchEngine: ContextualSearchEngine;
 
   constructor(
     private readonly app: App,
     private readonly settings: NovaJournalSettings
-  ) {}
+  ) {
+    this.searchEngine = new ContextualSearchEngine(settings);
+  }
 
   async incrementalUpdateIndex(folder: string): Promise<void> {
     if (!this.settings.aiEnabled || !this.settings.aiApiKey) return;
@@ -69,7 +75,7 @@ export class EnhancedEmbeddingService {
 
       for (const f of files) {
         const stat = this.app.vault.getAbstractFileByPath(f.path) as TFile;
-        const fileDate = this.extractDateFromFilename(f.name) || Date.now();
+        const fileDate = TemporalUtils.extractDateFromFilename(f.name) || Date.now();
         
         if (fileDate < cutoff) continue;
 
@@ -114,85 +120,31 @@ export class EnhancedEmbeddingService {
     k: number, 
     options: SearchOptions = {}
   ): Promise<EnhancedIndexedChunk[]> {
-    try {
-      await this.ensureIndexLoaded();
-      if (!this.index || this.index.items.length === 0) return [];
-
-      const { embeddings } = await embed({
-        apiKey: this.settings.aiApiKey,
-        inputs: [query.trim()],
-      });
-
-      if (!embeddings || embeddings.length === 0) return [];
-
-      const queryVector = embeddings[0];
-      let candidates = this.index.items.filter(item => 
-        item.vector && item.vector.length > 0
-      );
-
-      candidates = this.applyContextFilters(candidates, options);
-
-      const scored = candidates
-        .map((item) => ({
-          item,
-          score: this.calculateEnhancedScore(queryVector, item, query, options),
-        }))
-        .sort((a, b) => b.score - a.score);
-
-      const results = this.applyDiversityFilter(scored, options.diversityThreshold || 0.3);
-      
-      return results
-        .slice(0, Math.max(0, k))
-        .map((s) => ({
-          ...s.item,
-          text: `[${this.formatDate(s.item.date)}] ${s.item.text}`,
-        }));
-    } catch {
-      return [];
-    }
+    await this.ensureIndexLoaded();
+    if (!this.index || this.index.items.length === 0) return [];
+    
+    return this.searchEngine.performContextualSearch(query, k, this.index.items, options);
   }
 
   async emotionalSearch(query: string, mood: Partial<MoodData>, k: number = 5): Promise<EnhancedIndexedChunk[]> {
-    const emotionalTags = mood.dominant_emotions || [];
-    const sentiment = mood.sentiment;
+    await this.ensureIndexLoaded();
+    if (!this.index || this.index.items.length === 0) return [];
     
-    return this.contextualSearch(query, k, {
-      contextTypes: ['emotional', 'general'],
-      emotionalFilter: emotionalTags,
-      boostRecent: sentiment?.includes('negative'),
-      diversityThreshold: 0.4
-    });
+    return this.searchEngine.emotionalSearch(query, mood, k, this.index.items);
   }
 
   async temporalSearch(query: string, timeFrame: 'recent' | 'week' | 'month', k: number = 5): Promise<EnhancedIndexedChunk[]> {
-    const now = new Date();
-    let start: Date;
+    await this.ensureIndexLoaded();
+    if (!this.index || this.index.items.length === 0) return [];
     
-    switch (timeFrame) {
-      case 'recent':
-        start = new Date(now.getTime() - 3 * 24 * 60 * 60 * 1000);
-        break;
-      case 'week':
-        start = new Date(now.getTime() - 7 * 24 * 60 * 60 * 1000);
-        break;
-      case 'month':
-        start = new Date(now.getTime() - 30 * 24 * 60 * 60 * 1000);
-        break;
-    }
-
-    return this.contextualSearch(query, k, {
-      contextTypes: ['temporal', 'general'],
-      temporalRange: { start, end: now },
-      boostRecent: true
-    });
+    return this.searchEngine.temporalSearch(query, timeFrame, k, this.index.items);
   }
 
   async thematicSearch(query: string, themes: string[], k: number = 5): Promise<EnhancedIndexedChunk[]> {
-    return this.contextualSearch(query, k, {
-      contextTypes: ['thematic', 'general'],
-      thematicFilter: themes,
-      diversityThreshold: 0.5
-    });
+    await this.ensureIndexLoaded();
+    if (!this.index || this.index.items.length === 0) return [];
+    
+    return this.searchEngine.thematicSearch(query, themes, k, this.index.items);
   }
 
   private async fullRebuild(folder: string): Promise<void> {
@@ -211,7 +163,7 @@ export class EnhancedEmbeddingService {
 
     for (const file of files) {
       const stat = this.app.vault.getAbstractFileByPath(file.path) as TFile;
-      const fileDate = this.extractDateFromFilename(file.name) || Date.now();
+      const fileDate = TemporalUtils.extractDateFromFilename(file.name) || Date.now();
       
       if (fileDate < cutoff) continue;
       
@@ -264,10 +216,10 @@ export class EnhancedEmbeddingService {
   }
 
   private createEnhancedChunks(content: string, file: TFile): Omit<EnhancedIndexedChunk, 'vector'>[] {
-    const fileDate = this.extractDateFromFilename(file.name) || Date.now();
+    const fileDate = TemporalUtils.extractDateFromFilename(file.name) || Date.now();
     const chunks: Omit<EnhancedIndexedChunk, 'vector'>[] = [];
     
-    const textChunks = this.splitIntoChunks(content);
+    const textChunks = VectorUtils.splitIntoChunks(content);
     
     for (const text of textChunks) {
       if (text.trim().length < 50) continue;
@@ -277,11 +229,11 @@ export class EnhancedEmbeddingService {
         date: fileDate,
         lastModified: fileDate,
         text: text.trim(),
-        contextType: this.determineContextType(text),
-        emotionalTags: this.extractEmotionalTags(text),
-        thematicTags: this.extractThematicTags(text),
-        temporalMarkers: this.extractTemporalMarkers(text),
-        hash: this.hashString(text)
+        contextType: this.contextAnalyzer.determineContextType(text),
+        emotionalTags: this.contextAnalyzer.extractEmotionalTags(text),
+        thematicTags: this.contextAnalyzer.extractThematicTags(text),
+        temporalMarkers: this.contextAnalyzer.extractTemporalMarkers(text),
+        hash: VectorUtils.hashString(text)
       };
 
       chunks.push(chunk);
@@ -290,193 +242,7 @@ export class EnhancedEmbeddingService {
     return chunks;
   }
 
-  private determineContextType(text: string): ContextType {
-    const emotionalKeywords = /\b(feel|felt|emotion|mood|happy|sad|angry|frustrated|excited|anxious|calm|stressed|peaceful|worried|hopeful|disappointed|grateful|proud|embarrassed|confused|overwhelmed|content|joy|fear|love|hate|surprise|disgust|trust|anticipation)\b/i;
-    const temporalKeywords = /\b(today|yesterday|tomorrow|this week|last week|next week|this month|last month|recently|soon|now|then|when|during|after|before|while|since|until|ago|later)\b/i;
-    const thematicKeywords = /\b(work|job|career|family|friends|health|fitness|travel|hobby|project|goal|plan|study|learn|relationship|love|home|money|finance|food|exercise|book|movie|music|art|creative)\b/i;
 
-    let scores = {
-      emotional: 0,
-      temporal: 0,
-      thematic: 0
-    };
-
-    const emotionalMatches = text.match(emotionalKeywords);
-    const temporalMatches = text.match(temporalKeywords);
-    const thematicMatches = text.match(thematicKeywords);
-
-    if (emotionalMatches) scores.emotional = emotionalMatches.length;
-    if (temporalMatches) scores.temporal = temporalMatches.length;
-    if (thematicMatches) scores.thematic = thematicMatches.length;
-
-    const maxScore = Math.max(scores.emotional, scores.temporal, scores.thematic);
-    
-    if (maxScore === 0) return 'general';
-    
-    if (scores.emotional === maxScore) return 'emotional';
-    if (scores.temporal === maxScore) return 'temporal';
-    if (scores.thematic === maxScore) return 'thematic';
-    
-    return 'general';
-  }
-
-  private extractEmotionalTags(text: string): string[] {
-    const emotionMap: Record<string, string[]> = {
-      'positive': ['happy', 'excited', 'calm', 'peaceful', 'hopeful', 'grateful', 'proud', 'content', 'joy', 'love'],
-      'negative': ['sad', 'angry', 'frustrated', 'anxious', 'stressed', 'worried', 'disappointed', 'embarrassed', 'overwhelmed', 'fear', 'hate'],
-      'neutral': ['confused', 'surprised', 'curious', 'interested']
-    };
-
-    const found: string[] = [];
-    const lowerText = text.toLowerCase();
-
-    for (const [category, emotions] of Object.entries(emotionMap)) {
-      for (const emotion of emotions) {
-        if (lowerText.includes(emotion)) {
-          found.push(category);
-          break;
-        }
-      }
-    }
-
-    return [...new Set(found)];
-  }
-
-  private extractThematicTags(text: string): string[] {
-    const themeMap: Record<string, string[]> = {
-      'work': ['work', 'job', 'career', 'office', 'meeting', 'project', 'colleague', 'boss', 'deadline'],
-      'personal': ['family', 'friends', 'relationship', 'love', 'home', 'personal'],
-      'health': ['health', 'fitness', 'exercise', 'doctor', 'wellness', 'sleep', 'tired'],
-      'learning': ['study', 'learn', 'book', 'course', 'education', 'knowledge', 'skill'],
-      'creativity': ['art', 'creative', 'music', 'write', 'paint', 'design', 'create'],
-      'leisure': ['hobby', 'travel', 'movie', 'game', 'fun', 'vacation', 'relax']
-    };
-
-    const found: string[] = [];
-    const lowerText = text.toLowerCase();
-
-    for (const [theme, keywords] of Object.entries(themeMap)) {
-      for (const keyword of keywords) {
-        if (lowerText.includes(keyword)) {
-          found.push(theme);
-          break;
-        }
-      }
-    }
-
-    return [...new Set(found)];
-  }
-
-  private extractTemporalMarkers(text: string): string[] {
-    const timePatterns = [
-      /\b(today|yesterday|tomorrow)\b/gi,
-      /\b(this|last|next)\s+(week|month|year)\b/gi,
-      /\b(\d{1,2}:\d{2})\b/g,
-      /\b(morning|afternoon|evening|night)\b/gi,
-      /\b(\d{1,2})\s+(days?|weeks?|months?)\s+ago\b/gi
-    ];
-
-    const found: string[] = [];
-    
-    for (const pattern of timePatterns) {
-      const matches = text.match(pattern);
-      if (matches) {
-        found.push(...matches.map(m => m.toLowerCase()));
-      }
-    }
-
-    return [...new Set(found)];
-  }
-
-  private applyContextFilters(chunks: EnhancedIndexedChunk[], options: SearchOptions): EnhancedIndexedChunk[] {
-    let filtered = chunks;
-
-    if (options.contextTypes && options.contextTypes.length > 0) {
-      filtered = filtered.filter(chunk => 
-        options.contextTypes!.includes(chunk.contextType)
-      );
-    }
-
-    if (options.emotionalFilter && options.emotionalFilter.length > 0) {
-      filtered = filtered.filter(chunk =>
-        chunk.emotionalTags?.some(tag => 
-          options.emotionalFilter!.includes(tag)
-        )
-      );
-    }
-
-    if (options.thematicFilter && options.thematicFilter.length > 0) {
-      filtered = filtered.filter(chunk =>
-        chunk.thematicTags?.some(tag => 
-          options.thematicFilter!.includes(tag)
-        )
-      );
-    }
-
-    if (options.temporalRange) {
-      const { start, end } = options.temporalRange;
-      filtered = filtered.filter(chunk =>
-        chunk.date >= start.getTime() && chunk.date <= end.getTime()
-      );
-    }
-
-    return filtered;
-  }
-
-  private calculateEnhancedScore(
-    queryVector: number[], 
-    chunk: EnhancedIndexedChunk, 
-    query: string,
-    options: SearchOptions
-  ): number {
-    let baseScore = this.cosineSimilarity(queryVector, chunk.vector);
-
-    if (options.boostRecent) {
-      const ageInDays = (Date.now() - chunk.date) / (1000 * 60 * 60 * 24);
-      const recencyBoost = Math.exp(-ageInDays / 7);
-      baseScore *= (1 + recencyBoost * 0.2);
-    }
-
-    const queryLower = query.toLowerCase();
-    const textLower = chunk.text.toLowerCase();
-    
-    if (textLower.includes(queryLower)) {
-      baseScore *= 1.3;
-    }
-
-    if (chunk.contextType !== 'general') {
-      baseScore *= 1.1;
-    }
-
-    return baseScore;
-  }
-
-  private applyDiversityFilter(
-    scored: Array<{ item: EnhancedIndexedChunk; score: number }>, 
-    threshold: number
-  ): Array<{ item: EnhancedIndexedChunk; score: number }> {
-    if (threshold <= 0) return scored;
-
-    const result: Array<{ item: EnhancedIndexedChunk; score: number }> = [];
-    
-    for (const candidate of scored) {
-      let shouldInclude = true;
-      
-      for (const existing of result) {
-        const similarity = this.cosineSimilarity(candidate.item.vector, existing.item.vector);
-        if (similarity > threshold) {
-          shouldInclude = false;
-          break;
-        }
-      }
-      
-      if (shouldInclude) {
-        result.push(candidate);
-      }
-    }
-
-    return result;
-  }
 
   private removeChunksForFiles(filePaths: string[]): void {
     if (!this.index) return;
@@ -493,80 +259,12 @@ export class EnhancedEmbeddingService {
   private async computeFileHash(file: TFile): Promise<string> {
     const content = await this.app.vault.read(file);
     const mtime = file.stat?.mtime || 0;
-    return this.hashString(content + mtime.toString());
+    return VectorUtils.hashString(content + mtime.toString());
   }
 
-  private hashString(str: string): string {
-    let hash = 0;
-    for (let i = 0; i < str.length; i++) {
-      const char = str.charCodeAt(i);
-      hash = ((hash << 5) - hash) + char;
-      hash = hash & hash;
-    }
-    return hash.toString();
-  }
 
-  private splitIntoChunks(content: string): string[] {
-    const tokens = content.split(/\s+/);
-    const chunks: string[] = [];
-    let i = 0;
-    
-    while (i < tokens.length) {
-      const part = tokens
-        .slice(i, i + this.chunkSize)
-        .join(" ")
-        .trim();
-      if (part) chunks.push(part);
-      if (i + this.chunkSize >= tokens.length) break;
-      i += this.chunkSize - this.overlap;
-      if (i < 0) i = 0;
-    }
-    return chunks;
-  }
 
-  private cosineSimilarity(a: number[], b: number[]): number {
-    let dot = 0, na = 0, nb = 0;
-    const len = Math.min(a.length, b.length);
-    
-    for (let i = 0; i < len; i += 1) {
-      dot += a[i] * b[i];
-      na += a[i] * a[i];
-      nb += b[i] * b[i];
-    }
-    
-    if (na === 0 || nb === 0) return 0;
-    return dot / (Math.sqrt(na) * Math.sqrt(nb));
-  }
 
-  private extractDateFromFilename(filename: string): number | null {
-    // Support format YYYY-MM-DD_HH-ss or YYYY-MM-DD
-    const dateMatch = filename.match(/(\d{4}-\d{2}-\d{2})(?:_(\d{2})-(\d{2}))?/);
-    if (!dateMatch) return null;
-
-    const datePart = dateMatch[1];
-    const hourPart = dateMatch[2] || '00';
-    const minutePart = dateMatch[3] || '00';
-    
-    const date = new Date(`${datePart}T${hourPart}:${minutePart}:00`);
-    return isNaN(date.getTime()) ? null : date.getTime();
-  }
-
-  private formatDate(timestamp: number): string {
-    const date = new Date(timestamp);
-    const now = new Date();
-    const diffDays = Math.floor(
-      (now.getTime() - date.getTime()) / (1000 * 60 * 60 * 24)
-    );
-
-    if (diffDays === 0) return "aujourd'hui";
-    if (diffDays === 1) return "hier";
-    if (diffDays < 7) return `il y a ${diffDays} jours`;
-    if (diffDays < 30) return `il y a ${Math.floor(diffDays / 7)} semaines`;
-    return date.toLocaleDateString("fr-FR", {
-      month: "long",
-      day: "numeric",
-    });
-  }
 
   private getMarkdownFilesInFolder(folder: string): TFile[] {
     const files: TFile[] = [];
